+++ tests/test_entry.py
@@ -1,10 +1,8 @@
 """ Tests involved in CRUD for transaction Entries """
 import pytest
 from web import create_app, db
+from web.models import Entry, Setting, Character
 from config import TestConfig
-from web.entry.models import Entry
-from web.setting.models import Setting
-from web.character.models import Character
 
 
 @pytest.fixture(scope='session')
