+++ tests/test_setting.py
@@ -3,9 +3,8 @@ import pytest
 
 from config import TestConfig
 from web import db, create_app
+from web.models import Character, Setting
 from web.setting.utility import get_setting, save_setting
-from web.character.models import Character
-from web.setting.models import Setting
 
 
 @pytest.fixture(scope='session')
