+++ web/__init__.py
@@ -3,6 +3,7 @@ from flask_bootstrap import Bootstrap
 from config import DevConfig
 from flask_sqlalchemy import SQLAlchemy
 from flask_migrate import Migrate
+# from os import path
 
 # create global objects
 bs = Bootstrap()
@@ -33,5 +34,11 @@ def create_app():
 
     from web.entry.routes import entry_bp
     app.register_blueprint(entry_bp)
+
+    # if not path.exists('.\\web\\wallet.sqlite'):
+    #     print('Path not Found: Call Create')
+    #     db.create_all()
+    # else:
+    #     print('Path Found')
 
     return app
