+++ web/entry/routes.py
@@ -1,10 +1,9 @@
 from flask import Blueprint, render_template, redirect, url_for, request
 
 from web import db
+from web.models import Entry, Character
 from web.entry.forms import AddEntryForm, EditEntryForm
-from web.entry.models import Entry
 from web.entry.utility import get_current_character_id, get_balance, get_game_session_list
-from web.character.models import Character
 from web.setting.utility import get_setting, save_setting
 
 entry_bp = Blueprint('entry_bp', __name__, template_folder='templates')
