+++ web/entry/utility.py
@@ -1,8 +1,7 @@
-from web.entry.models import Entry
-from web.setting.models import Setting
-from web.character.models import Character
-from web import db
 from flask import session
+
+from web import db
+from web.models import Entry, Character, Setting
 
 
 def get_balance():
