+++ web/models/__init__.py
@@ -1,0 +1,34 @@
+from web import db
+from sqlalchemy import Column, Integer, ForeignKey
+from sqlalchemy.orm import relationship
+
+
+class Entry(db.Model):
+    """ Represents an entry in the purse """
+    __tablename__ = 'Entries'
+    id = db.Column(db.Integer, primary_key=True)
+    game_session = db.Column(db.Integer)
+    description = db.Column(db.String(150), default='')
+    amount = db.Column(db.Float)
+    character_id = Column(Integer, ForeignKey('Characters.id'))
+
+
+class Character(db.Model):
+    """ Represents a character """
+    __tablename__ = 'Characters'
+    id = db.Column(db.Integer, primary_key=True)
+    name = db.Column(db.String(40), nullable=False)
+    is_dead = db.Column(db.Boolean, default=False)
+    entries = relationship('Entry', backref='character')
+
+
+class Setting(db.Model):
+    """ represents persistent settings for the application """
+    __tablename__ = 'Settings'
+    id = db.Column(db.Integer, primary_key=True)
+    key = db.Column(db.String(), nullable=False)
+    value = db.Column(db.String(), nullable=False)
+
+
+def init_db():
+    db.create_all()
